feat(routes): implement route protection - Add gate that calls isAdmin() helper function for accessing admin areas - Reference gate in middleware for admin-only routes - Pass is_admin property using HandleInertiaRequests share() method

+++ routes/web.php
@@ -17,16 +17,16 @@ Route::get('/contact', function() {
 
 // Index
 Route::get('/showtimes', [ShowtimeController::class, 'index']);
-Route::get('/showtimes/create', [ShowtimeController::class, 'create']);
-Route::get('/showtimes/{showtime}', [ShowtimeController::class, 'show']);
-Route::get('/showtimes/{showtime}/edit', [ShowtimeController::class, 'edit']);
-Route::patch('/showtimes/{showtime}', [ShowtimeController::class, 'update']);
-Route::delete('/showtimes/{showtime}', [ShowtimeController::class, 'destroy']);
-Route::post('/showtimes', [ShowtimeController::class, 'store']);
+Route::get('/showtimes/create', [ShowtimeController::class, 'create'])->middleware(['auth', 'can:access-admin-area']);
+Route::get('/showtimes/{showtime}', [ShowtimeController::class, 'show'])->middleware(['auth', 'can:access-admin-area']);
+Route::get('/showtimes/{showtime}/edit', [ShowtimeController::class, 'edit'])->middleware(['auth', 'can:access-admin-area']);
+Route::patch('/showtimes/{showtime}', [ShowtimeController::class, 'update'])->middleware(['auth', 'can:access-admin-area']);
+Route::delete('/showtimes/{showtime}', [ShowtimeController::class, 'destroy'])->middleware(['auth', 'can:access-admin-area']);
+Route::post('/showtimes', [ShowtimeController::class, 'store'])->middleware(['auth', 'can:access-admin-area']);
 
 Route::get('/register', [RegisteredUserController::class, 'create']);
 Route::post('/register', [RegisteredUserController::class, 'store']);
 
-Route::get('/login', [SessionController::class, 'create']);
+Route::get('/login', [SessionController::class, 'create'])->name('login');
 Route::post('/login', [SessionController::class, 'store']);
 Route::post('/logout', [SessionController::class, 'destroy']);
